[FIX] activity PDP styles
